[BUILD] In update pipeline always try to write summary (#35)
## Changes
The package update pipeline can fail if there's an issue with one of the packages. ~~I believe there's some logging into the generated MD file~~ It's good to have the md output even if it doesn't bring any new info on failure, so we need to consume it in the next step instead of cancelling the pipeline.

The actual approach to mitigate those issues is to disable the restore in the pipeline. Instead it will create the PR and then we will observe the failure on the PR build (if it exists at that point). I suspect the issue is that restoring projects after each package change can fail while restoring them all at once after updates would succeed in many cases.

* [BUILD] In update pipeline always try to write summary
* Disable Restore in update pipeline