Add OTLP export from test server (#40)
## Changes
OTEL in .NET supports configuration through env variables to point at a collector. I'm exposing those variables in the docker-compose file and setting them during deployment in GitHub Action.

## Motivation
I want to be able to see the telemetry emitted by the server in the test deployment. The GitHub repo is set up with details of my Honeycomb account.